e2e: framework: clean/export NUMA alignment utils
We recently added topology manager e2e tests.
This pulled in some utilities to check the NUMA alignment
of the resources of a container, and these utilities can and
should be generalized and exported in the e2e framework.

Doing this will enable us to do further cleanup and consolidation
in this area, and make it easier to add more tests for both existing
components and for future extensions which are on the roadmap in the
future (example: memory manager).
