+++ test/e2e_node/BUILD
@@ -15,7 +15,6 @@ go_library(
         "framework.go",
         "image_list.go",
         "node_problem_detector_linux.go",
-        "numa_alignment.go",
         "resource_collector.go",
         "util.go",
         "util_sriov.go",
@@ -33,7 +32,6 @@ go_library(
         "//pkg/kubelet/apis/podresources/v1alpha1:go_default_library",
         "//pkg/kubelet/apis/stats/v1alpha1:go_default_library",
         "//pkg/kubelet/cm:go_default_library",
-        "//pkg/kubelet/cm/cpuset:go_default_library",
         "//pkg/kubelet/cri/remote:go_default_library",
         "//pkg/kubelet/kubeletconfig/util/codec:go_default_library",
         "//pkg/kubelet/metrics:go_default_library",
@@ -207,6 +205,7 @@ go_test(
         "//test/e2e/framework/skipper:go_default_library",
         "//test/e2e/framework/testfiles:go_default_library",
         "//test/e2e/framework/volume:go_default_library",
+        "//test/e2e_node/alignment:go_default_library",
         "//test/e2e_node/perf/workloads:go_default_library",
         "//test/e2e_node/services:go_default_library",
         "//test/utils:go_default_library",
@@ -266,6 +265,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
+        "//test/e2e_node/alignment:all-srcs",
         "//test/e2e_node/builder:all-srcs",
         "//test/e2e_node/environment:all-srcs",
         "//test/e2e_node/perf/workloads:all-srcs",
